+++ app/src/main/java/org/ieee/ieeehub/provider/conferencetag/ConferenceTagModel.java
@@ -17,7 +17,7 @@ public interface ConferenceTagModel extends BaseModel {
      * Can be {@code null}.
      */
     @Nullable
-    Integer getConferenceId();
+    Long getConferenceId();
 
     /**
      * Get the {@code name} value.
